% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp1919-package.R
\docType{package}
\name{cp1919-package}
\alias{cp1919}
\alias{cp1919-package}
\title{cp1919: Radio Observations of the Pulse Profiles and Dispersion Measures of Twelve Pulsars}
\description{
Successive pulses from CP 1919, the first pulsar discovered. The pulses occur every 1.337 seconds. They are caused by rapidly spinning neutron star.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cp1919}
  \item Report bugs at \url{https://github.com/pachadotdev/cp1919/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
\keyword{internal}
